requirements.txt
ultralytics==8.3.6
opencv-python-headless==4.10.0.84
streamlit==1.39.0
numpy==1.26.4
Pillow==10.4.0
